pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-discord-bot"
version = "1.0-dev"
requires-python = ">=3.12"

dependencies = [
    "aiohttp==3.10.5",
    "discord.py==2.4.0",
    "python-dotenv==1.0.1",
    "yt-dlp==2024.12.23",
    "jellyfin-apiclient-python==1.10.0",
    "validators==0.34.0",
    "google-api-python-client==2.158.0",
]

[[tool.mypy.overrides]]
module = [
    "jellyfin_apiclient_python.*",
    "googleapiclient.*",
]
follow_untyped_imports = true

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "locally-disabled",
    "missing-class-docstring",
    "missing-module-docstring",
    "missing-function-docstring",
]
